refactor(state): cleanup store/context logic for budget and auth (#211)
- Remove duplicate authStore.js and consolidate to AuthContext.jsx
- Remove unnecessary BudgetContext.jsx wrapper and useBudget.js hook
- Update all components to use direct useBudgetStore access
- Remove unused useBudgetQuery.js hook with overlapping logic
- Add missing setEncryption and updateProfile methods to AuthContext
- Fix useAuthFlow, usePasswordRotation, useDataManagement to use unified auth
- Standardize state management to use Zustand stores directly
- Remove redundant Context API layers for cleaner architecture

✅ All auth and budget flows tested and working
✅ ESLint passing with reduced warnings (20 vs previous 25+)
✅ No unused state or dead code remaining
✅ Hooks follow consistent pattern for accessing store state
